- EthereumBlobExporter refactoring for relay and container as a source
