+++ chains/orchestrator-relays/runtime/dancelight/src/xcm_config.rs
@@ -46,7 +46,6 @@ use {
     sp_runtime::traits::TryConvertInto,
     tanssi_runtime_common::relay::ExporterFeeHandler,
     tp_bridge::{
-        container_token_to_ethereum_message_exporter::ContainerEthereumBlobExporter,
         snowbridge_outbound_token_transfer::{EthereumBlobExporter, SnowbrigeTokenTransferRouter},
         EthereumLocationsConverterFor,
     },
@@ -256,7 +255,7 @@ impl xcm_executor::Config for XcmConfig {
         WaivedLocations,
         ExporterFeeHandler<Self::AssetTransactor, SnowbridgeFeesAccount, TreasuryAccount>,
     >;
-    type MessageExporter = (SnowbridgeExporter, ContainerToSnowbridgeMessageExporter);
+    type MessageExporter = SnowbridgeExporter;
     type UniversalAliases = Nothing;
     type CallDispatcher = RuntimeCall;
     type SafeCallFilter = Everything;
@@ -402,16 +401,6 @@ parameter_types! {
 pub type SnowbridgeExporter = EthereumBlobExporter<
     UniversalLocation,
     EthereumNetwork,
-    snowbridge_pallet_outbound_queue::Pallet<Runtime>,
-    EthereumSystem,
-    SnowbridgeChannelInfo,
->;
-
-/// Exports message to the Ethereum Gateway contract.
-pub type ContainerToSnowbridgeMessageExporter = ContainerEthereumBlobExporter<
-    UniversalLocation,
-    EthereumNetwork,
-    EthereumLocation,
     snowbridge_pallet_outbound_queue::Pallet<Runtime>,
     EthereumSystem,
     SnowbridgeChannelInfo,
